+++ lib/stored_session/gem_version.rb
@@ -8,7 +8,7 @@ module StoredSession
 
   module VERSION
     MAJOR = 0
-    MINOR = 2
+    MINOR = 3
     TINY  = 0
     PRE   = "".freeze