+++ CHANGELOG.md
@@ -1,10 +1,10 @@
-# Stored Session Changelog
+# Solid Session Changelog
 
 ## Unreleased
 
 ## [0.2.0](https://github.com/tbhb/stored_session/releases/tag/v0.2.0)
 
-- Rename gem to Stored Session
+- Rename gem to `stored_session` (<https://github.com/tbhb/stored_session/pull/19>)
 
 ## [0.1.0](https://github.com/tbhb/stored_session/releases/tag/v0.1.0)