+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add log subscriber events (<https://github.com/tbhb/stored_session/pull/22>)
+
 ## [0.2.0](https://github.com/tbhb/stored_session/releases/tag/v0.2.0)
 
 - Rename gem to `stored_session` (<https://github.com/tbhb/stored_session/pull/19>)