+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add log subscriber events (<https://github.com/tbhb/stored_session/pull/22>)
+- Extract session model behavior into concern (<https://github.com/tbhb/stored_session/pull/24>)
 
 ## [0.2.0](https://github.com/tbhb/stored_session/releases/tag/v0.2.0)