+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [0.3.0](https://github.com/tbhb/stored_session/releases/tag/v0.3.0)
+
 - Add log subscriber events (<https://github.com/tbhb/stored_session/pull/22>)
 - Extract session model behavior into concern (<https://github.com/tbhb/stored_session/pull/24>)
 - Rename trim to expire and add Rake tasks (<https://github.com/tbhb/stored_session/pull/27>)